fixed bug, downloading git scripts works now

+++ install_on_macos.sh
@@ -33,19 +33,24 @@ chsh -s $bash
 sudo chsh -s $bash
 
 echo "Downloading scripts used for the git status in the prompt ..."
+sudo mkdir -p /etc/bash_completion.d
 sudo curl -o /etc/bash_completion.d/git-completion https://raw.githubusercontent.com/git/git/master/contrib/completion/git-completion.bash
 sudo curl -o /etc/bash_completion.d/git-prompt https://raw.githubusercontent.com/git/git/master/contrib/completion/git-prompt.sh
 
 echo "Downlaoding custom .bashrc ..."
 curl -o ~/.bashrc https://raw.githubusercontent.com/ValentinKolb/bashrc/main/.bashrc
 
-read -p "Do you want also want to replace the '~/.profile' file? [Y/n] " -n 1 -r
-echo
-if [[ $REPLY =~ ^[Yy]$ ]] || [[ -z $REPLY ]]; then
-    rm ~/.profile
-    curl -o ~/.profile https://raw.githubusercontent.com/ValentinKolb/bashrc/main/.profile
+if [[ -f "~/.profile" ]]; then
+    read -p "Do you want also want to replace the '~/.profile' file? [Y/n] " -n 1 -r
+    echo
+    if [[ $REPLY =~ ^[Yy]$ ]] || [[ -z $REPLY ]]; then
+        rm ~/.profile
+        curl -o ~/.profile https://raw.githubusercontent.com/ValentinKolb/bashrc/main/.profile
+    else
+        echo "Ok. Please make sure to load the ~/.bashrc in your ~/.profile file."
+    fi
 else
-    echo "Ok. Please make sure to load the ~/.bashrc in your ~/.profile file."
+    curl -o ~/.profile https://raw.githubusercontent.com/ValentinKolb/bashrc/main/.profile
 fi
 
 echo "Instalation complete. Please restart your shell now."
